+++ python-lib/dku_aws/eksctl_command.py
@@ -32,8 +32,9 @@ class EksctlCommand(object):
     def run_and_get_output(self):
         result = self.run()
         if result[1] != 0:
+            logging.error(result[3])
             cmd = _convert_to_string([self.eksctl_bin] + self.args)
-            raise Exception('Failed to execute command: \'%s\'.' % (' '.join(cmd)))
+            raise Exception('Failed to execute command: \'%s\'. See log for more details.' % (' '.join(cmd)))
         else:
             return result[2]