+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version 1.0.9 - Feature and bugfix release
-- Throwing an exception when command launched by `EksctlCommand.run_and_get_output` or `AwsCommand.run_and_get_output` fails
+- Throwing an exception when command invoked by `EksctlCommand.run_and_get_output` or `AwsCommand.run_and_get_output` fails
 
 ## Version 1.0.8 - Feature and bugfix release
 - Add option to install Metrics Server